clean code a bit
